Implemented selection methods
The ```utils/selections.py``` file containts multiple classes with a common interface for selecting members from a provided population.

A number of these contain a *preprocess* step which reduces the amount of computation required in a lot of cases.

This commit also adds the TravisCI badge to the ```README.md```