Bounds now given in dictionary form
Altered ```Individual``` to make bounds given in dictionary form - this allows the user to provide labels for each of the bounds which can be very useful.

In addition the ```README.md``` has been updated to reflect this change.

All unit tests have been refactored to reflect this change as well.

Resolves: #26